# new: Significantly speed up gk for small epsilon