Speed up gk for small epsilons by computing compression bands on the fly
rather than precomputing them.  This results in an approximate 25 times
speedup for epsilon = 0.1.

Before:

```
2021-12-22T14:49:23-06:00
Running /home/sengelha/.cache/bazel/_bazel_sengelha/96bd37d8c4611ae46e087160c1a47334/execroot/__main__/bazel-out/k8-opt/bin/cpp/benchmark
Run on (8 X 3096 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x1)
Load Average: 0.65, 0.53, 0.35
--------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------
gk e=0.1 n=1000000 sorted         18789311886 ns   18789108600 ns            1 Rate=53.2223k/s
gk e=0.01 n=1000000 sorted         241657813 ns    241656933 ns            3 Rate=1.37937M/s
gk e=0.001 n=1000000 sorted        458908319 ns    458906950 ns            2 Rate=1089.55k/s
```

After:

```
2021-12-22T14:50:18-06:00
Running /home/sengelha/.cache/bazel/_bazel_sengelha/96bd37d8c4611ae46e087160c1a47334/execroot/__main__/bazel-out/k8-opt/bin/cpp/benchmark
Run on (8 X 3096 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x1)
Load Average: 0.74, 0.58, 0.38
--------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------
gk e=0.1 n=1000000 sorted           17992271 ns     17992321 ns           38 Rate=1.46261M/s
gk e=0.01 n=1000000 sorted          66318772 ns     66318518 ns           11 Rate=1.37079M/s
gk e=0.001 n=1000000 sorted        477852821 ns    477853300 ns            2 Rate=1046.35k/s
```